+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - **NEW**: new op: `SGN` (sign of argument)
 - **NEW**: new kria remote op: `KR.DUR`
 - **NEW**: new op: `NR` (alias #) (binary math pattern generator)
-- **NEW**: new ops: `SN, CN, CSN` (use western scales and chords to get values from `N` table)
+- **NEW**: new ops: `N.S, N.C, N.CS` (use western scales and chords to get values from `N` table)
 - **FIX**: dim M in edit mode when metro inactive
 
 ## v3.1.0